AgenticRAG/src/requirements.txt
langchain==0.3.19
langchain-cohere==0.4.2
python-dotenv==1.0.1
langchain-chroma==0.2.2